.github/scripts/bootstrap/requirements.txt
PyGithub==2.6.1
rich==14.0.0
cryptography==45.0.3
